requirements.txt
pandas>=1.3.0
numpy>=1.20.0
requests>=2.25.0
tqdm>=4.60.0
openpyxl>=3.0.7
pyyaml>=6.0.0
openai>=1.0.0
anthropic>=0.3.0
python-dotenv>=0.19.0
tenacity>=8.0.0 